requirements.txt
matplotlib==3.1.1
numpy==1.17.0
Pillow==6.0.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.9.0
torch==1.1.0
torch-spline-conv==1.1.0
torchvision==0.3.0
